--- a/packages/runtime-core/src/componentProxy.ts
+++ b/packages/runtime-core/src/componentProxy.ts
@@ -201,7 +201,6 @@ const PublicInstanceProxyHandlers: ProxyHandler<any> = {
           value
         })
       } else {
-        console.log('setting', key)
         ctx[key] = value
       }
     }