As mentioned in https://github.com/vuejs/vue-next/issues/1227, variable access from templates is a major hotspot in terms of performance. This PR reduces up to 80% of the variable access time. It does so by creating ultra-fast and optimizable getter functions.

The peformance difference can be seen here. Notice that the left side is with the changes in this PR, right side is the current Vue implementation. The 'get' time is much lower (70 vs 420) and overall it causes a performance improvement of about 10%.

![performance](https://user-images.githubusercontent.com/120531/83033513-7e402580-a037-11ea-89ae-b963a01b7fd0.png)

In terms of memory usage, getter functions must be created per instance. Initial tests proved that it seems to has little effect on the memory usage.