@basvanmeurs 2020-05-28
Please don't merge this pull request yet. I'd like to do some further analysis on the performance of creating new component instances, memory usage and optimizer behavior.