@basvanmeurs 2020-05-29
I changed some things in the original idea and did some extensive performance testing, comparing the current vue-next version with this PR. The results are further below, but most importantly they show that instance proxy performance in component instance creation has become about 15% slower (because some functions and function arguments need to be put into the propGetters cache), while setup access (particles test) has become about 75% faster. It only makes a slight impact on the total time, although in some cases (when there are many v-for items or components, and many setup accesses), there is certainly a performance improvement with this PR.

In general, nothing beats exposing the setup state (as raw object) directly and changing the syntax to `setup.myvar.value` and `props.myprop` instead of accepting just `myvar` and `props` in the template. This gives the best possible performance, makes it clear immediately where the property is coming from and is closest to writing your own render functions.

Nevertheless, I realise that we want to preserve the existing syntax. So for that case this PR can help limiting the performance difference somewhat. Still, I would prefer to see 'setup' and 'prop' being directly available, as mentioned by Robbin.

100 new components per frame, 200 frames
Previous flushJobs: 1194ms, 1236ms, 1213ms
Previous proxy get: 287ms, 262m, 308ms
New flushJobs: 1115ms, 1171ms, 1161ms
New proxy get: 325ms, 335ms, 316ms

100 reused components per frame, 200 frames
Previous flushJobs: 862ms, 1038ms, 720ms
Previous proxy get: 155ms, 175ms, 135ms
New flushJobs: 786ms, 991ms, 937ms
New proxy get: 136ms, 101ms, 124ms

Particles test (6000 updated v-for items within the same component, all requesting 1 setup ref):
Previous flushJobs: 
1905, 1893, 1864
Previous proxy get:
376, 405, 410

New flushJobs: 
1759, 1643, 1725
New proxy get:
88, 103, 102