@yyx990803 2020-06-25
First of all, thanks a lot of coming up with the idea and experimenting with this. Unfortunately, the gain here doesn't seem to justify the increased complexity of the implementation.

I still think this is an area worth exploring - but I think a more compiler-informed approach would be more plausible.